Moved function and added dry-run
